refactor(tray): implement tray icon management service and update service imports
